Add deploy sub command

+++ justfile
@@ -1,8 +1,14 @@
 # Build cleaned data with elevation and population
 build-data:
-    rm -f public/data/*
-    cargo run -r -p data-cli  -- build-admin1
-    cargo run -r -p data-cli  -- build-data
+    #!/usr/bin/env bash
+    set -e
+    if [ -f public/data/admin1.json.br ] && [ -f public/data/cities.jsonl.br ]; then
+        echo "Data files already exist, skipping regeneration"
+    else
+        rm -f public/data/*
+        cargo run -r -p data-cli -- build-admin1
+        cargo run -r -p data-cli -- build-data
+    fi
 
 # Build WASM package
 build-wasm:
@@ -21,3 +27,8 @@ build-zip: build-web build-data
         cd public && \
         zip -r ../bundle.zip ./
 
+deploy: build-zip
+    cargo run -r -p data-cli -- \
+        deploy-cf-pages \
+        --account-id $(op --account my.1password.ca read op://Personal/zman.mendy.dev/account_id) \
+        --token $(op --account my.1password.ca read op://Personal/zman.mendy.dev/api_token)
